+++ src/git_gerrit/action/DLCherryPickAction.py
@@ -1,7 +1,7 @@
 import urwid
-from Git import Git
-from Button import Button
-import ErrorPopup
+from git_gerrit.model.Git import Git
+from git_gerrit.model.Button import Button
+from git_gerrit.model.ErrorPopup import display
 
 class DLCherryPickAction(urwid.WidgetWrap):
     def __init__(self, chageview):
@@ -13,6 +13,6 @@ class DLCherryPickAction(urwid.WidgetWrap):
         try:
             git.fetch(self.cview.change['revisions'][self.cview.active_revision_sha]['ref'])
             git.cherry_pick(self.cview.active_revision_sha)
-            ErrorPopup.display(self.cview.main, "Cherry-Pick successful")
+            display(self.cview.main, "Cherry-Pick successful")
         except Exception as e:
-            ErrorPopup.display(self.cview.main, str(e))
+            display(self.cview.main, str(e))
