+++ src/git_gerrit/action/GitResetPrevAction.py
@@ -1,8 +1,8 @@
 import urwid
 
-from Git import Git
-from Button import Button
-import ErrorPopup
+from git_gerrit.model.Git import Git
+from git_gerrit.model.Button import Button
+from git_gerrit.model.ErrorPopup import display
 
 class GitResetPrevAction(urwid.WidgetWrap):
     def __init__(self, chageview):
@@ -15,4 +15,4 @@ class GitResetPrevAction(urwid.WidgetWrap):
             git.fetch(self.cview.change['revisions'][self.cview.active_revision_sha]['ref'])
             git.reset("HEAD~1")
         except Exception as e:
-            ErrorPopup.display(self.cview.main, str(e))
+            display(self.cview.main, str(e))
