+++ src/git_gerrit/action/ManageReviewersAction.py
@@ -2,8 +2,8 @@
 
 import urwid
 
-from Button import Button
-from InputHandler import InputHandler
+from git_gerrit.model.Button import Button
+from git_gerrit.model.InputHandler import InputHandler
 
 class ManageReviewersAction(urwid.WidgetWrap):
     def __init__(self, chageview):
