+++ src/git_gerrit/model/CommitMsg.py
