+++ src/git_gerrit/model/InputHandler.py
@@ -1,6 +1,6 @@
 import urwid
 
-from Button import Button
+from git_gerrit.model.Button import Button
 
 class InputHandler(urwid.WidgetWrap):
     def __init__(self, widget, callbacks):
