+++ src/git_gerrit/model/SelectableListItem.py
